策略接口联调，修复系统模块BUG

+++ hdoc-system/src/main/java/com/insigma/ordercenter/controller/SysRoleController.java
@@ -8,8 +8,13 @@ import com.insigma.ordercenter.base.BaseRequest;
 import com.insigma.ordercenter.base.CodeMsg;
 import com.insigma.ordercenter.base.Result;
 import com.insigma.ordercenter.entity.SysRole;
+import com.insigma.ordercenter.entity.SysUser;
+import com.insigma.ordercenter.entity.UserRoleRelation;
 import com.insigma.ordercenter.entity.vo.SysRoleVO;
 import com.insigma.ordercenter.service.ISysRoleService;
+import com.insigma.ordercenter.service.ISysUserService;
+import com.insigma.ordercenter.service.IUserRoleRelationService;
+import com.insigma.ordercenter.service.IUserShopRelationService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
@@ -32,10 +37,16 @@ import java.util.List;
 @RestController
 @RequestMapping("/sys-role")
 @Slf4j
-public class SysRoleController {
+public class SysRoleController extends BaseController{
 
     @Resource
     private ISysRoleService sysRoleService;
+
+    @Resource
+    private ISysUserService iSysUserService;
+
+    @Resource
+    private IUserRoleRelationService iUserRoleRelationService;
 
     @PutMapping("/add")
     @ApiOperation("新增角色")
@@ -46,6 +57,7 @@ public class SysRoleController {
         if (count > 0) {
             return Result.success(CodeMsg.ROLE_NAME_REPEAT);
         }
+        data.setCreateId(redisUser().getUserId());
         data.setCreateTime(LocalDateTime.now());
         boolean status = sysRoleService.save(data);
         if (status) {
@@ -58,6 +70,16 @@ public class SysRoleController {
     @DeleteMapping("/delete/{Id}")
     @ApiOperation("删除角色")
     public Result delete(@PathVariable Long Id) {
+
+        // 判断角色是否绑定账号，绑定则不允许删除
+        QueryWrapper<UserRoleRelation> wrapper = new QueryWrapper<>();
+        wrapper.eq(UserRoleRelation.ROLE_ID,Id);
+        int count = this.iUserRoleRelationService.count(wrapper);
+
+        if(0 != count){
+            return Result.error(CodeMsg.ROLE_USER);
+        }
+
         boolean status = sysRoleService.removeById(Id);
         if (status) {
             return Result.success();
@@ -88,6 +110,12 @@ public class SysRoleController {
     @ApiOperation("获取角色列表")
     public Result list() {
         List<SysRole> result = sysRoleService.list();
+
+        result.forEach(sysRole -> {
+
+            SysUser user = this.iSysUserService.getById(sysRole.getCreateId());
+            sysRole.setCreateName(user.getUserName());
+        });
         return Result.success(result);
     }
